+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 【core】NetUtil增加isInRange方法（感谢@【成都】小邓）
 * 【core】新增BeanPath
 * 【core】CollUtil新增reverse、reverseNew方法
+* 【core】集合中新增排列（Arrangement）和组合（Combination）类
 
 ### Bug修复
 * 【core】修复IdcardUtil中身份证15转18位年的问题（Issue#IHT1Q@Gitee）